+++ Sources/KeyboardKit/Keyboard/KeyboardInputViewController.swift
@@ -6,6 +6,7 @@
 //  Copyright © 2021 Daniel Saidi. All rights reserved.
 //
 
+import SwiftUI
 import UIKit
 
 /**
@@ -51,6 +52,36 @@ open class KeyboardInputViewController: UIInputViewController {
     open override func traitCollectionDidChange(_ previousTraitCollection: UITraitCollection?) {
         keyboardContext.sync(with: self)
         super.traitCollectionDidChange(previousTraitCollection)
+        trySetupLastView()
+    }
+    
+    
+    // MARK: - Setup
+    
+    private var lastSetupView: AnyView?
+    
+    /**
+     Remove all subviews from the vc view and add a `SwiftUI`
+     view to it. This will pin the view to the edges of this
+     extension and will resize the extension to fit the view.
+     
+     This function also applies `@EnvironmentObject`s to the
+     view, that can be used by all nested views.
+     */
+    open func setup<Content: View>(with view: Content) {
+        lastSetupView = AnyView(view)
+        trySetupLastView()
+    }
+    
+    private func trySetupLastView() {
+        guard let view = lastSetupView else { return }
+        self.view.subviews.forEach { $0.removeFromSuperview() }
+        let hostingView = view
+            .environmentObject(keyboardContext)
+            .environmentObject(keyboardInputCalloutContext)
+            .environmentObject(keyboardSecondaryInputCalloutContext)
+        let controller = KeyboardHostingController(rootView: hostingView)
+        controller.add(to: self)
     }
     
     
@@ -107,18 +138,6 @@ open class KeyboardInputViewController: UIInputViewController {
         inputSetProvider: keyboardInputSetProvider)
     
     /**
-     Setting this property will set the context locale, then
-     call `setupKeyboard` to ensure that it changes.
-     */
-    public var keyboardLocale: Locale {
-        get { keyboardContext.locale }
-        set {
-            primaryLanguage = newValue.languageCode
-            keyboardContext.locale = newValue
-        }
-    }
-    
-    /**
      The extension's default secondary input callout context.
      */
     public lazy var keyboardSecondaryInputActionProvider: SecondaryCalloutActionProvider = StandardSecondaryCalloutActionProvider(
@@ -130,15 +149,6 @@ open class KeyboardInputViewController: UIInputViewController {
     public lazy var keyboardSecondaryInputCalloutContext = SecondaryInputCalloutContext(
         actionProvider: keyboardSecondaryInputActionProvider,
         actionHandler: keyboardActionHandler)
-    
-    /**
-     Setting this property will update the context type then
-     call `setupKeyboard` to ensure that it changes.
-     */
-    public var keyboardType: KeyboardType {
-        get { keyboardContext.keyboardType }
-        set { keyboardContext.keyboardType = newValue }
-    }
     
     
     // MARK: - View Properties
@@ -157,7 +167,7 @@ open class KeyboardInputViewController: UIInputViewController {
     }()
     
     
-    // MARK: - Keyboard Functionality
+    // MARK: - Text And Selection Change
     
     open override func selectionWillChange(_ textInput: UITextInput?) {
         super.selectionWillChange(textInput)
@@ -169,40 +179,6 @@ open class KeyboardInputViewController: UIInputViewController {
         resetAutocomplete()
     }
     
-    
-    // MARK: - Public Functions
-    
-    /**
-     Change the keyboard locale.
-     
-     Although you can just set the `keyboardLocale` property
-     directly, this func is injected to the action handler.
-     */
-    open func changeKeyboardLocale(to locale: Locale) {
-        keyboardLocale = locale
-    }
-    
-    /**
-     Change the keyboard type.
-     
-     Although you can just set `keyboardType` directly, this
-     func is injected into the action handler.
-     */
-    open func changeKeyboardType(to type: KeyboardType) {
-        keyboardType = type
-    }
-    
-    
-    // MARK: - Autocomplete
-    
-    
-    open func performAutocomplete() {}
-    
-    open func resetAutocomplete() {}
-    
-    
-    // MARK: - UITextInputDelegate
-    
     open override func textWillChange(_ textInput: UITextInput?) {
         super.textWillChange(textInput)
         keyboardContext.textDocumentProxy = textDocumentProxy
@@ -213,6 +189,41 @@ open class KeyboardInputViewController: UIInputViewController {
         performAutocomplete()
         tryChangeToPreferredKeyboardTypeAfterTextDidChange()
     }
+    
+    
+    // MARK: - Public Functions
+    
+    /**
+     Change the keyboard locale.
+     
+     This function can be called directly or injected into a
+     nested service class to avoid bilinear dependencies.
+     */
+    open func changeKeyboardLocale(to locale: Locale) {
+        primaryLanguage = locale.languageCode
+        keyboardContext.locale = locale
+    }
+    
+    /**
+     Change the keyboard type.
+     
+     This function can be called directly or injected into a
+     nested service class to avoid bilinear dependencies.
+     */
+    open func changeKeyboardType(to type: KeyboardType) {
+        keyboardContext.keyboardType = type
+    }
+    
+    /**
+     Perform an autocomplete operation. You can override the
+     function to provide custom autocomplete logic.
+     
+     This function can be called directly or injected into a
+     nested service class to avoid bilinear dependencies.
+     */
+    open func performAutocomplete() {}
+    
+    open func resetAutocomplete() {}
 }