+++ Sources/KeyboardKit/System/SystemKeyboardItem.swift
@@ -38,7 +38,6 @@ public struct SystemKeyboardItem<Content: View>: View {
     
     public var body: some View {
         content
-            .frame(maxWidth: .infinity)
             .frame(height: item.size.height - item.insets.top - item.insets.bottom)
             .width(item.size.width, totalWidth: keyboardSize.width, referenceSize: referenceSize)
             .keyboardButtonStyle(for: item.action, appearance: appearance)